Replace `Spot` with `Position`, a simple object that Firebase can handle
