+++ app/src/main/java/ch/epfl/sweng/eventmanager/repository/data/EventLocation.java
@@ -10,19 +10,15 @@ public class EventLocation {
      * The public name for the location (e.g. EPFL)
      */
     private String name;
-    /**
-     * The latitude of the location
-     */
-    private double latitude;
-    /**
-     * The longitude of the location
-     */
-    private double longitude;
 
-    public EventLocation(String name, double latitude, double longitude) {
+    /**
+     * The actual coordinates of the location
+     */
+    private Position position;
+
+    public EventLocation(String name, Position position) {
         this.name = name;
-        this.latitude = latitude;
-        this.longitude = longitude;
+        this.position = position;
     }
 
     public EventLocation() {
@@ -32,11 +28,7 @@ public class EventLocation {
         return name;
     }
 
-    public double getLatitude() {
-        return latitude;
-    }
-
-    public double getLongitude() {
-        return longitude;
+    public Position getPosition() {
+        return position;
     }
 }
